Add files via upload
In recent times  the  concept  of  smart  cities have  gained  grate  popularity.Smart car parking system is one of them. 

In this project, we have implemented a smart parking system.In which we have given access of the authentic cars. Without authentication, no car will enter. If all slots of the system is full, the new authentic car which is waiting to enter will get red signal. There are some pictures which can describe the project.